Move another server-only utility
