+++ packages/ai/src/chains/copywriter/schema.ts
@@ -1,5 +1,4 @@
 import { z } from "zod";
-import type { Instance, Instances } from "@webstudio-is/sdk";
 
 /**
  * Copywriter chain.
@@ -33,63 +32,3 @@ export type Context = z.infer<typeof ContextSchema>;
 
 export const ResponseSchema = z.array(TextInstanceSchema);
 export type Response = z.infer<typeof ResponseSchema>;
-
-export const collectTextInstances = ({
-  instances,
-  rootInstanceId,
-  textComponents = new Set(["Heading", "Paragraph", "Text"]),
-}: {
-  instances: Instances;
-  rootInstanceId: Instance["id"];
-  textComponents?: Set<string>;
-}) => {
-  const textInstances: TextInstance[] = [];
-
-  const rootInstance = instances.get(rootInstanceId);
-
-  if (rootInstance === undefined) {
-    return textInstances;
-  }
-
-  const nodeType =
-    rootInstance.component === "Heading" ||
-    rootInstance.component === "Paragraph"
-      ? rootInstance.component
-      : "Text";
-
-  // Instances can have a number of text child nodes without interleaving components.
-  // When this is the case we treat the child nodes as a single text node,
-  // otherwise the AI would generate children.length chunks of separate text.
-  // To signal that a textInstance is "joint" we set the index to -1.
-  if (rootInstance.children.every((child) => child.type === "text")) {
-    textInstances.push({
-      instanceId: rootInstanceId,
-      index: -1,
-      type: nodeType,
-      text: rootInstance.children.map((child) => child.value).join(" "),
-    });
-  } else {
-    rootInstance.children.forEach((child, index) => {
-      if (child.type === "text") {
-        if (textComponents.has(rootInstance.component)) {
-          textInstances.push({
-            instanceId: rootInstanceId,
-            index,
-            type: nodeType,
-            text: child.value,
-          });
-        }
-      } else if (child.type === "id") {
-        textInstances.push(
-          ...collectTextInstances({
-            instances,
-            rootInstanceId: child.value,
-            textComponents,
-          })
-        );
-      }
-    });
-  }
-
-  return textInstances;
-};
